Changes cache and api data to LiveData.

+++ app/src/main/java/com/example/nearby_finder/PlacesViewModel.kt
@@ -11,20 +11,23 @@ import com.example.nearby_finder.data.PlaceItem
 import com.example.nearby_finder.data.PlacesRepository
 import com.example.nearby_finder.managers.NetworkManager
 import com.example.nearby_finder.managers.PlaceManager
+import com.google.android.libraries.places.api.model.Place
+import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.launch
 import kotlin.concurrent.thread
 
 class PlacesViewModel(private val repository: PlacesRepository): ViewModel() {
 
-    var places: LiveData<List<PlaceItem>> = PlaceManager.getFetchedPlacesAsFlow().asLiveData()
+    var places = MutableLiveData<List<PlaceItem>>()
 
     private val bubbleSort = BubbleSort()
 
     private fun insertAll() = viewModelScope.launch {
-        repository.deleteAll()
-        val list = bubbleSort.sortList(PlaceManager.getFetchedPlaces())
-        repository.insertAll(list)
+        if (PlaceManager.list.value != null) {
+            repository.deleteAll()
+            repository.insertAll(PlaceManager.list.value as MutableList<PlaceItem>)
+        }
     }
 
     class PlacesViewModelFactory(private val repository: PlacesRepository) : ViewModelProvider.Factory {
@@ -49,14 +52,14 @@ class PlacesViewModel(private val repository: PlacesRepository): ViewModel() {
             val connectivityManager = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
             connectivityManager.registerDefaultNetworkCallback(object : ConnectivityManager.NetworkCallback() {
                 override fun onAvailable(network: Network?) {
-                    places = PlaceManager.getFetchedPlacesAsFlow().asLiveData()
+                    places = repository.places
                 }
 
                 override fun onLost(network: Network?) {
-                    if (PlaceManager.getFetchedPlacesAsFlow() != repository.places) {
+                    if (PlaceManager.list.value != null) {
                         insertAll()
                     }
-                    places = repository.places.asLiveData()
+                    places = repository.places
                 }
             }
             )
